LPS-71164 Move methods that are specific to a specific SourceCheck over there instead of passing it to constructor
