+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaOSGiReferenceCheck.java
@@ -30,6 +30,8 @@ import com.liferay.source.formatter.util.FileUtil;
 
 import java.io.File;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -43,12 +45,10 @@ import java.util.regex.Pattern;
  */
 public class JavaOSGiReferenceCheck extends BaseFileCheck {
 
-	public JavaOSGiReferenceCheck(
-		Map<String, String> moduleFileNamesMap,
-		List<String> serviceReferenceUtilClassNames) {
-
-		_moduleFileNamesMap = moduleFileNamesMap;
-		_serviceReferenceUtilClassNames = serviceReferenceUtilClassNames;
+	public JavaOSGiReferenceCheck() throws Exception {
+		_moduleFileNamesMap = _getModuleFileNamesMap();
+		_serviceReferenceUtilClassNames = getPropertyList(
+			"service.reference.util.class.names");
 	}
 
 	@Override
@@ -424,6 +424,44 @@ public class JavaOSGiReferenceCheck extends BaseFileCheck {
 		int pos = serviceDirLocation.lastIndexOf(".com.");
 
 		return serviceDirLocation.substring(pos + 1);
+	}
+
+	private Map<String, String> _getModuleFileNamesMap() throws Exception {
+		Map<String, String> moduleFileNamesMap = new HashMap<>();
+
+		List<String> fileNames = new ArrayList<>();
+
+		String moduleRootDirLocation = "modules/";
+
+		for (int i = 0; i < 6; i++) {
+			File file = new File(getBaseDirName() + moduleRootDirLocation);
+
+			if (file.exists()) {
+				fileNames = getFileNames(
+					getBaseDirName() + moduleRootDirLocation,
+					new String[0], new String[] {"**/*.java"});
+
+				break;
+			}
+
+			moduleRootDirLocation = "../" + moduleRootDirLocation;
+		}
+
+		for (String fileName : fileNames) {
+			fileName = StringUtil.replace(
+				fileName, CharPool.BACK_SLASH, CharPool.SLASH);
+
+			String className = StringUtil.replace(
+				fileName, CharPool.SLASH, CharPool.PERIOD);
+
+			int pos = className.lastIndexOf(".com.liferay.");
+
+			className = className.substring(pos + 1, fileName.length() - 5);
+
+			moduleFileNamesMap.put(className, fileName);
+		}
+
+		return moduleFileNamesMap;
 	}
 
 	private String _getModuleSuperClassContent(
