+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaTermOrderCheck.java
@@ -15,6 +15,7 @@
 package com.liferay.source.formatter.checks;
 
 import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.tools.ToolsUtil;
 import com.liferay.source.formatter.parser.JavaClass;
 import com.liferay.source.formatter.parser.JavaStaticBlock;
 import com.liferay.source.formatter.parser.JavaTerm;
@@ -24,14 +25,16 @@ import com.liferay.source.formatter.util.FileUtil;
 import java.io.File;
 
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
  * @author Hugo Huijser
  */
 public class JavaTermOrderCheck extends BaseJavaTermCheck {
 
-	public JavaTermOrderCheck(String portalCustomSQLContent) {
-		_portalCustomSQLContent = portalCustomSQLContent;
+	public JavaTermOrderCheck() throws Exception {
+		_portalCustomSQLContent = _getPortalCustomSQLContent();
 	}
 
 	@Override
@@ -62,6 +65,32 @@ public class JavaTermOrderCheck extends BaseJavaTermCheck {
 
 	protected String[] getCheckableJavaTermNames() {
 		return new String[] {JAVA_CLASS};
+	}
+
+	private String _getPortalCustomSQLContent() throws Exception {
+		if (!isPortalSource()) {
+			return null;
+		}
+
+		File portalCustomSQLFile = getFile(
+			"portal-impl/src/custom-sql/default.xml",
+			ToolsUtil.PORTAL_MAX_DIR_LEVEL);
+
+		String portalCustomSQLContent = FileUtil.read(portalCustomSQLFile);
+
+		Matcher matcher = _customSQLFilePattern.matcher(portalCustomSQLContent);
+
+		while (matcher.find()) {
+			File customSQLFile = getFile(
+				"portal-impl/src/" + matcher.group(1),
+				ToolsUtil.PORTAL_MAX_DIR_LEVEL);
+
+			if (customSQLFile != null) {
+				portalCustomSQLContent += FileUtil.read(customSQLFile);
+			}
+		}
+
+		return portalCustomSQLContent;
 	}
 
 	private String _getCustomSQLContent(String fileName, String absolutePath)
@@ -158,6 +187,8 @@ public class JavaTermOrderCheck extends BaseJavaTermCheck {
 		return javaClass.getContent();
 	}
 
+	private final Pattern _customSQLFilePattern = Pattern.compile(
+		"<sql file=\"(.*)\" \\/>");
 	private final String _portalCustomSQLContent;
 
 }
