+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaVariableTypeCheck.java
@@ -14,6 +14,9 @@
 
 package com.liferay.source.formatter.checks;
 
+import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.portal.kernel.util.MapUtil;
+import com.liferay.portal.kernel.util.SetUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -35,14 +38,10 @@ import java.util.regex.Pattern;
  */
 public class JavaVariableTypeCheck extends BaseJavaTermCheck {
 
-	public JavaVariableTypeCheck(
-		List<String> annotationsExclusions,
-		Map<String, String> defaultPrimitiveValues,
-		Set<String> immutableFieldTypes) {
-
-		_annotationsExclusions = annotationsExclusions;
-		_defaultPrimitiveValues = defaultPrimitiveValues;
-		_immutableFieldTypes = immutableFieldTypes;
+	public JavaVariableTypeCheck() {
+		_annotationsExclusions = _getAnnotationsExclusions();
+		_defaultPrimitiveValues = _getDefaultPrimitiveValues();
+		_immutableFieldTypes = _getImmutableFieldTypes();
 	}
 
 	@Override
@@ -239,6 +238,23 @@ public class JavaVariableTypeCheck extends BaseJavaTermCheck {
 		return childJavaTerms;
 	}
 
+	private List<String> _getAnnotationsExclusions() {
+		return ListUtil.fromArray(
+			new String[] {
+				"ArquillianResource", "Autowired", "BeanReference", "Captor",
+				"Inject", "Mock", "Parameter", "Reference", "ServiceReference",
+				"SuppressWarnings", "Value"
+			});
+	}
+
+	private Map<String, String> _getDefaultPrimitiveValues() {
+		return MapUtil.fromArray(
+			new String[] {
+				"boolean", "false", "char", "'\\\\0'", "byte", "0", "double",
+				"0\\.0", "float", "0\\.0", "int", "0", "long", "0", "short", "0"
+			});
+	}
+
 	private String _getFieldType(JavaVariable javaVariable) {
 		StringBundler sb = new StringBundler(4);
 
@@ -256,6 +272,19 @@ public class JavaVariableTypeCheck extends BaseJavaTermCheck {
 		}
 
 		return null;
+	}
+
+	private Set<String> _getImmutableFieldTypes() {
+		Set<String> immutableFieldTypes = SetUtil.fromArray(
+			new String[] {
+				"boolean", "byte", "char", "double", "float", "int", "long",
+				"short", "Boolean", "Byte", "Character", "Class", "Double",
+				"Float", "Int", "Long", "Number", "Short", "String"
+			});
+
+		immutableFieldTypes.addAll(getPropertyList("immutable.field.types"));
+
+		return immutableFieldTypes;
 	}
 
 	private boolean _isFinalableField(
